pyproject.toml
[project]
name = "mtntripdata"
version = "0.1.0"
description = "Build a database of mountaineer activities by scraping data from mountaineer.org"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg[binary,pool]>=3.2.6",
    "python-dotenv>=1.0.1",
    "selenium>=4.29.0",
    "sqlalchemy>=2.0.39",
]
